REG: buggy new function comitted and quick fixed

+++ testbeam_analysis/dut_alignment.py
@@ -244,9 +244,15 @@ def merge_cluster_data(cluster_files, alignment_file, tracklets_file, pixel_size
     p = np.cumsum(common_bin_count)
     i = np.zeros(p[-1], np.int)
     np.add.at(i, p[:-1], 1)
-    common_event_number = np.cumsum(i)
+    common_event_number = np.cumsum(i).astype(np.int64)
     # this does the same:
     # common_event_number = np.repeat(np.arange(common_bin_count.size), common_bin_count)
+
+#     # Calculate a event number index to map the cluster of all files to
+#     common_event_number_2 = None
+#     for cluster_file in cluster_files:
+#         with tb.open_file(cluster_file, mode='r') as in_file_h5:
+#             common_event_number_2 = in_file_h5.root.Cluster[:]['event_number'] if common_event_number_2 is None else analysis_utils.get_max_events_in_both_arrays(common_event_number_2, in_file_h5.root.Cluster[:]['event_number'])
 
     # Create result array description, depends on the number of DUTs
     description = [('event_number', np.int64)]
